# Optimize queries that need to sort on `recent_crate_downloads.downloads`