@jtgeibel 2019-07-14
> The crate still won't show up in search until the view gets
updated though, which occurs every 10 minutes.

Its probably fairly common for a user to search for their crate shortly after publishing for the first time.  This is also a bit further reaching, as the dashboard relies on the search endpoint.  It hits `crates?user_id=123` and `crates?following=1`. (The list of crates on team pages are also populated by search results from this endpoint.)

Especially for search results and the dashboard, I'm afraid a delay of up to 10 minutes in results could confuse users.

> Two other options would be to add the index without the join change/view
definition change, which would mean that the summary page gets fixed,
but search is still slow. Summary is much more frequently hit than
search ordered by recent downloads, so this wouldn't be the end of the
world.

I think this alternative sounds like a solid improvement that avoids the potential for confusing search results in various places in the frontend.