@smarnach 2019-08-19
Once we provide database dumps, the load from crawlers will likely reduce over time as well.